Moved all stuff concernig Claculations into one module. SPlit different Calculation types into different classes.
