+++ flixOpt/core.py
@@ -142,113 +142,6 @@ class TimeSeries:
             self.explicit_active_data = self.make_scalar_if_possible(explicit_active_data)
 
 
-class TimeSeriesCollection:
-    '''
-    calculates weights of TimeSeries for being in that collection (depending on)
-    '''
-
-    @property
-    def addPeak_Max_labels(self):
-        if self._addPeakMax_labels == []:
-            return None
-        else:
-            return self._addPeakMax_labels
-
-    @property
-    def addPeak_Min_labels(self):
-        if self._addPeakMin_labels == []:
-            return None
-        else:
-            return self._addPeakMin_labels
-
-    def __init__(self,
-                 time_series_list: List[TimeSeries],
-                 addPeakMax_TSraw: Optional[List[TimeSeriesRaw]] = None,
-                 addPeakMin_TSraw: Optional[List[TimeSeriesRaw]] = None):
-        self.time_series_list = time_series_list
-        self.addPeakMax_TSraw = addPeakMax_TSraw or []
-        self.addPeakMin_TSraw = addPeakMin_TSraw or []
-        # i.g.: self.agg_type_count = {'solar': 3, 'price_el' = 2}
-        self.agg_type_count = self._get_agg_type_count()
-
-        self._checkPeak_TSraw(addPeakMax_TSraw)
-        self._checkPeak_TSraw(addPeakMin_TSraw)
-
-        # these 4 attributes are now filled:
-        self.seriesDict = {}
-        self.weightDict = {}
-        self._addPeakMax_labels = []
-        self._addPeakMin_labels = []
-        self.calculateParametersForTSAM()
-
-    def calculateParametersForTSAM(self):
-        for i in range(len(self.time_series_list)):
-            aTS: TimeSeries
-            aTS = self.time_series_list[i]
-            # check uniqueness of label:
-            if aTS.label_full in self.seriesDict.keys():
-                raise Exception('label of TS \'' + str(aTS.label_full) + '\' exists already!')
-            # add to dict:
-            self.seriesDict[
-                aTS.label_full] = aTS.active_data_vector  # Vektor zuweisen!# TODO: müsste doch active_data sein, damit abhängig von Auswahlzeitraum, oder???
-            self.weightDict[aTS.label_full] = self._getWeight(aTS)  # Wichtung ermitteln!
-            if (aTS.TSraw is not None):
-                if aTS.TSraw in self.addPeakMax_TSraw:
-                    self._addPeakMax_labels.append(aTS.label_full)
-                if aTS.TSraw in self.addPeakMin_TSraw:
-                    self._addPeakMin_labels.append(aTS.label_full)
-
-    def _get_agg_type_count(self):
-        # count agg_types:
-        from collections import Counter
-
-        TSlistWithAggType = []
-        for TS in self.time_series_list:
-            if self._get_agg_type(TS) is not None:
-                TSlistWithAggType.append(TS)
-        agg_types = (aTS.TSraw.agg_group for aTS in TSlistWithAggType)
-        return Counter(agg_types)
-
-    def _get_agg_type(self, aTS: TimeSeries):
-        if (aTS.TSraw is not None):
-            agg_type = aTS.TSraw.agg_group
-        else:
-            agg_type = None
-        return agg_type
-
-    def _getWeight(self, aTS: TimeSeries):
-        if aTS.TSraw is None:
-            # default:
-            weight = 1
-        elif aTS.TSraw.agg_weight is not None:
-            # explicit:
-            weight = aTS.TSraw.agg_weight
-        elif aTS.TSraw.agg_group is not None:
-            # via agg_group:
-            # i.g. n=3 -> weight=1/3
-            weight = 1 / self.agg_type_count[aTS.TSraw.agg_group]
-        else:
-            weight = 1
-            # raise Exception('TSraw is without weight definition.')
-        return weight
-
-    def _checkPeak_TSraw(self, aTSrawlist):
-        if aTSrawlist is not None:
-            for aTSraw in aTSrawlist:
-                if not isinstance(aTSraw, TimeSeriesRaw):
-                    raise Exception('addPeak_max/min must be list of TimeSeriesRaw-objects!')
-
-    def print(self):
-        print('used ' + str(len(self.time_series_list)) + ' TS for aggregation:')
-        for TS in self.time_series_list:
-            aStr = ' ->' + TS.label_full + ' (weight: {:.4f}; agg_group: ' + str(self._get_agg_type(TS)) + ')'
-            print(aStr.format(self._getWeight(TS)))
-        if len(self.agg_type_count.keys()) > 0:
-            print('agg_types: ' + str(list(self.agg_type_count.keys())))
-        else:
-            print('Warning!: no agg_types defined, i.e. all TS have weigth 1 (or explicit given weight)!')
-
-
 def as_effect_dict(effect_values: Union[Numeric, TimeSeries, Dict]) -> Optional[Dict]:
     """
     Converts effect values into a dictionary. If a scalar value is provided, it is associated with a standard effect type.
