+++ flixOpt/modeling.py
@@ -429,46 +429,6 @@ class VariableTS(Variable):
         return value, time
 
 
-# managed die Before-Werte des segments:
-class BeforeValues:
-    def __init__(self,
-                 variables_ts: List[VariableTS],
-                 lastUsedIndex: int):
-        self.beforeValues = {}
-        # Sieht dann so aus = {(Element1, aVar1.name): (value, time),
-        #                      (Element2, aVar2.name): (value, time),
-        #                       ...                       }
-        for aVar in variables_ts:
-            aVar: VariableTS
-            if aVar.activated_beforeValues:
-                # Before-Value holen:
-                (aValue, aTime) = aVar.get_before_value_for_next_segment(lastUsedIndex)
-                self.addBeforeValues(aVar, aValue, aTime)
-
-    def addBeforeValues(self, aVar, aValue, aTime):
-        element = aVar.owner
-        aKey = (element, aVar.label)  # hier muss label genommen werden, da aVar sich ja ändert je linear_model!
-        # before_values = aVar.result(aValue) # letzten zwei Werte
-
-        if aKey in self.beforeValues.keys():
-            raise Exception('setBeforeValues(): Achtung Wert würde überschrieben, Wert ist schon belegt!')
-        else:
-            self.beforeValues.update({aKey: (aValue, aTime)})
-
-    # return (value, time)
-    def getBeforeValues(self, aVar):
-        element = aVar.owner
-        aKey = (element, aVar.label)  # hier muss label genommen werden, da aVar sich ja ändert je linear_model!
-        if aKey in self.beforeValues.keys():
-            return self.beforeValues[aKey]  # returns (value, time)
-        else:
-            return None
-
-    def print(self):
-        for (element, varName) in self.beforeValues.keys():
-            print(element.label + '__' + varName + ' = ' + str(self.beforeValues[(element, varName)]))
-
-
 # class cInequation(Equation):
 #   def __init__(self, label, owner, linear_model):
 #     super().__init__(label, owner, linear_model, eqType='ineq')
